--- nMap.py
@@ -1,2 +1,0 @@
-def nMap(n, min1, max1, min2, max2):
-    return ((n-min1)/(max1-min1)) * (max2-min2) + min2
