--- python.py
@@ -1,9 +1,0 @@
-from ursina import *
-# Petter Amland
-from ursina.prefabs.first_person_controller import FirstPersonController
-from numpy import floor, abs, sin, cos, radians
-# from random import randrange
-import time
-from perlin_noise import PerlinNoise
-from nMap import nMap
-from cave_system import Caves
